Short explainer: created directories for individual poker games, changed up namespaces, added TCP
I moved what was previously the interpreter for Trey into Five (Not "FCP" so as not to
be confused with four card poker, oft shortened to FCP) and cloned FCP and modified it
to interpret three card poker hands. I kept in the ability to have a board, however I
made it optional, and am looking to make it optional in Five as well. TCP's board also
has no length restriction nor cap.